Bug 590389 – plparse/tests/parser is broken
Change how signals are emitted from the playlist parsing process so they're
once more emitted synchronously when we're parsing synchronously, and in
idle callbacks when we're parsing asynchronously. This solves the awkward
problem of all the TotemPlParser::entry-parsed signals being emitted after
a call to a synchronous parse function had returned. Closes: bgo#590389