Make sure videos disguised as MP4 aren't unhandled
*.mp4 files were ignored if they were actually detected
as any other type. This was the case for 3gpp files, but
it would happen for *all* files types that weren't listed.

And we can't list every file type as possibly being a quicktime file.

So we ignore files that are text/plain, or a sub-class of it.

This makes FLV hiding as an MP4, or 3GPP hiding as an MP4 work
correctly.

https://bugzilla.gnome.org/show_bug.cgi?id=620039