introspection: Skip TotemPlParserMetadata because it’s evil
TotemPlParserMetadata is a GHashTable wearing a fake moustache, originating
from the days when GHashTable wasn’t a proper boxed type. GIR knows about the
peculiarities of GHashTable (for example, its key and value element types);
but is fooled by the fake moustache and doesn’t know what to make of
TotemPlParserMetadata.

This is bad. Nobody likes fake moustaches.

This commit marks TotemPlParserMetadata as to be (skip)ped by GIR, and
overrides the locations where it used to pretend to be GHashTables. Since
the copy/free functions for the boxed TotemPlParserMetadata and boxed
GHashTable are the same, this shouldn’t cause problems for bindings when
they try and unbox the GHashTables, since no boxed type identifier is carried
along with the box.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=669398