build: Use gnome-common code coverage macros to add gcov support
This replaces totem-pl-parser’s existing home-grown solution with the
gnome-common code coverage macros added in gnome-common 3.5.6/3.6.0.

To use:
    cd plparse && make check-code-coverage
this will run the totem-pl-parser test suite and build a code coverage
report of the code which was touched. The report will be an HTML file and
its URI will be printed on completion.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=683212