plparse: Handle length attribute for M3U playlists
When an M3U playlist advertises the length of an item as -1, we
don't need to check whether it is another playlist, it would usually
be a file of unknown duration, or a stream.

This matches the current code in the PLS parser, and works around
hang-on-close() bugs in gvfs.

https://bugzilla.gnome.org/show_bug.cgi?id=695652