plparse: Fix parsing of local HLS playlists
Local HLS playlists will have their mime-type already set to the one
gathered through magic because of its locality. As the
application/vnd.apple.mpegurl mime-type descends from text/plain, it
will be ignored (as it shouldn't be needed to go through the M3U parser
for it) instead of unhandled.

Fix this, and one of our tests, by checking for that playlist type.