plparse: Port from GSimpleAsyncResult to GTask
GSimpleAsyncResult was deprecated in version 2.36 of GLib, and replaced
by GTask. Bump our GLib dependency to 2.36 and start using GTask instead
to avoid deprecated function warnings from the compiler.

Furthermore, as we are targetting GNOME 3.10 and later versions, and
GNOME 3.10 shipped with GLib 2.38, we don't need to worry about
leaving older versions in the dust.

https://bugzilla.gnome.org/show_bug.cgi?id=776259