podcast: Always prefer feed tags with longer descriptions
This is to handle rss feeds with shorter subtitle than description
(or vice-versa), where the shorter subtitle occurs after the longer
description overwriting the full description with a partial
subtitle. We now always look for the longer description of
<description>, <itunes:summary> and <itunes:subtitle> for feeds.

Note that <description> contains either the full HTML (or a shortened
version if a <content:encoded> element is available, <itunes:summary> is
a plain-text <description> and <itunes:subtitle> shortened, plain-text
<description>, so <description> should always be getting selected over
those shorter versions.