tests: Fix invalid UTF-8 RSS test
The test expected us to pass invalid UTF-8 through, and just feed it to
the calling application.

GLib APIs actually guarantee that data passed is UTF-8 unless stated
otherwise, so make sure that no data is parsed if the data isn't UTF-8
and we can't detect what encoding it's in.