core: fix garbage in CDATA values for shorter texts
This change properly null terminates the CDATA values. For shorter
CDATA values, this fix is mandatory to override the token buffer text
which contains the previous token value of '<![CDATA['. CDATA value is
overwritten on top of this CDATA marker causing garbage data.

For longer CDATA values, this wasn't an issue, as the token buffer (
64K ) is already filled with NULL chars, causing this issue to not
show up.

Closes: #30